refactor: remove unnecessary dependencies
