+++ app/component/EmptyFavouriteLocationSlot.js
@@ -2,9 +2,6 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import Link from 'found/lib/Link';
 import { FormattedMessage } from 'react-intl';
-// import AddIcon from './AddIcon';
-import ComponentUsageExample from './ComponentUsageExample';
-import { addAnalyticsEvent } from '../util/analyticsUtils';
 
 const EmptyFavouriteLocationSlot = ({ index }) => {
   return (
@@ -13,13 +10,6 @@ const EmptyFavouriteLocationSlot = ({ index }) => {
       to="/suosikki/uusi"
       className="cursor-pointer no-decoration"
       key={`add-new-favourite-${index}`}
-      onClick={() => {
-        addAnalyticsEvent({
-          category: 'Favourite',
-          action: 'AddLocationAsFavourite',
-          name: null,
-        });
-      }}
     >
       <div className="new-favourite-button-content">
         <p className="add-location-text">
@@ -31,15 +21,6 @@ const EmptyFavouriteLocationSlot = ({ index }) => {
 };
 
 EmptyFavouriteLocationSlot.displayName = 'EmptyFavouriteLocationSlot';
-
-EmptyFavouriteLocationSlot.description = () => (
-  <div>
-    <p>Renders a empty favourite location slot component</p>
-    <ComponentUsageExample description="none">
-      <EmptyFavouriteLocationSlot />
-    </ComponentUsageExample>
-  </div>
-);
 
 EmptyFavouriteLocationSlot.propTypes = {
   index: PropTypes.number.isRequired,
